correction

+++ src/Controller/ClientController.php
@@ -16,9 +16,14 @@ class ClientController extends AbstractController
     #[Route('/', name: 'app_client_index', methods: ['GET'])]
     public function index(ClientRepository $clientRepository): Response
     {
-        return $this->render('client/index.html.twig', [
-            'clients' => $clientRepository->findAll(),
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('client/index.html.twig', [
+                'clients' => $clientRepository->findAll(),
+            ]);
+        }
+        else {
+            return $this->redirectToRoute('app_accueil');
+        }
     }
 
     #[Route('/new', name: 'app_client_new', methods: ['GET', 'POST'])]
@@ -28,24 +33,34 @@ class ClientController extends AbstractController
         $form = $this->createForm(ClientType::class, $client);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid() && $this->isGranted('ROLE_ADMIN')) {
             $clientRepository->save($client, true);
 
             return $this->redirectToRoute('app_client_index', [], Response::HTTP_SEE_OTHER);
         }
 
+        if($this->isGranted('ROLE_ADMIN')){
         return $this->renderForm('client/new.html.twig', [
             'client' => $client,
             'form' => $form,
         ]);
     }
+        else {
+            return $this->redirectToRoute('app_accueil');
+        }
+    }
 
     #[Route('/{id}', name: 'app_client_show', methods: ['GET'])]
     public function show(Client $client): Response
     {
-        return $this->render('client/show.html.twig', [
-            'client' => $client,
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('client/show.html.twig', [
+                'client' => $client,
+            ]);
+        }
+        else {
+            return $this->redirectToRoute('app_accueil');
+        }
     }
 
     #[Route('/{id}/edit', name: 'app_client_edit', methods: ['GET', 'POST'])]
@@ -54,22 +69,27 @@ class ClientController extends AbstractController
         $form = $this->createForm(ClientType::class, $client);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid() && $this->isGranted('ROLE_ADMIN')) {
             $clientRepository->save($client, true);
 
             return $this->redirectToRoute('app_client_index', [], Response::HTTP_SEE_OTHER);
         }
 
+        if ($this->isGranted('ROLE_ADMIN')) {
         return $this->renderForm('client/edit.html.twig', [
             'client' => $client,
             'form' => $form,
         ]);
     }
+    else {
+        return $this->redirectToRoute('app_accueil');
+    }
+    }
 
     #[Route('/{id}', name: 'app_client_delete', methods: ['POST'])]
     public function delete(Request $request, Client $client, ClientRepository $clientRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$client->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete'.$client->getId(), $request->request->get('_token')) && $this->isGranted('ROLE_ADMIN')) {
             $clientRepository->remove($client, true);
         }
