+++ src/Controller/DevisCrudController.php
@@ -16,9 +16,14 @@ class DevisCrudController extends AbstractController
     #[Route('/', name: 'app_devis_crud_index', methods: ['GET'])]
     public function index(DevisRepository $devisRepository): Response
     {
-        return $this->render('devis_crud/index.html.twig', [
-            'devis' => $devisRepository->findAll(),
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('devis_crud/index.html.twig', [
+                'devis' => $devisRepository->findAll(),
+            ]);
+        }
+        else {
+            return $this->redirectToRoute('app_accueil');
+        }
     }
 
     #[Route('/new', name: 'app_devis_crud_new', methods: ['GET', 'POST'])]
@@ -28,24 +33,33 @@ class DevisCrudController extends AbstractController
         $form = $this->createForm(DevisType::class, $devi);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid() && $this->isGranted('ROLE_ADMIN')) {
             $devisRepository->save($devi, true);
 
             return $this->redirectToRoute('app_devis_crud_index', [], Response::HTTP_SEE_OTHER);
         }
-
+        if($this->isGranted('ROLE_ADMIN')){
         return $this->renderForm('devis_crud/new.html.twig', [
             'devi' => $devi,
             'form' => $form,
         ]);
     }
+        else {
+            return $this->redirectToRoute('app_accueil');
+        }
+    }
 
     #[Route('/{id}', name: 'app_devis_crud_show', methods: ['GET'])]
     public function show(Devis $devi): Response
     {
-        return $this->render('devis_crud/show.html.twig', [
-            'devi' => $devi,
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('devis_crud/show.html.twig', [
+                'devi' => $devi,
+            ]);
+        }
+        else {
+            return $this->redirectToRoute('app_accueil');
+        }
     }
 
     #[Route('/{id}/edit', name: 'app_devis_crud_edit', methods: ['GET', 'POST'])]
@@ -54,22 +68,27 @@ class DevisCrudController extends AbstractController
         $form = $this->createForm(DevisType::class, $devi);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+        if ($form->isSubmitted() && $form->isValid() && $this->isGranted('ROLE_ADMIN')) {
             $devisRepository->save($devi, true);
 
             return $this->redirectToRoute('app_devis_crud_index', [], Response::HTTP_SEE_OTHER);
         }
 
+        if($this->isGranted('ROLE_ADMIN')){
         return $this->renderForm('devis_crud/edit.html.twig', [
             'devi' => $devi,
             'form' => $form,
         ]);
+        }
+        else {
+            return $this->redirectToRoute('app_accueil');
+        }
     }
 
     #[Route('/{id}', name: 'app_devis_crud_delete', methods: ['POST'])]
     public function delete(Request $request, Devis $devi, DevisRepository $devisRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$devi->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete'.$devi->getId(), $request->request->get('_token')) && $this->isGranted('ROLE_ADMIN')) {
             $devisRepository->remove($devi, true);
         }
